tutorial/requirements.txt
pyspark==3.4.0
pre-commit==3.0.2
jupyter==1.0.0
kafka-python==2.0.2
